Saparate AutoSizeButton Gonna need a machete after this one
