+++ GamesToGo.Desktop/Graphics/ObjectManagerContainer.cs
@@ -20,7 +20,7 @@ namespace GamesToGo.Desktop.Graphics
         private IBindableList<ProjectElement> localElements = new BindableList<ProjectElement>();
 
         private FillFlowContainer<ElementEditButton> allElements;
-        private AddElementButton addElementButton;
+        private AutoSizeButton addElementButton;
         private readonly string title;
         private readonly string buttonText;
 
@@ -157,7 +157,7 @@ namespace GamesToGo.Desktop.Graphics
                                         RelativeSizeAxes = Axes.Both,
                                         Colour = Color4.Beige
                                     },
-                                    addElementButton = new AddElementButton
+                                    addElementButton = new AutoSizeButton
                                     {
                                         Action = () => project.AddElement(new T()),
                                         Text = buttonText
@@ -216,45 +216,6 @@ namespace GamesToGo.Desktop.Graphics
             }
 
             ItemsRemoved?.Invoke(resultingRemoved);
-        }
-
-
-        private class AddElementButton : GamesToGoButton
-        {
-            private readonly Container content = new Container
-            {
-                RelativeSizeAxes = Axes.Y,
-                AutoSizeAxes = Axes.X,
-            };
-
-            protected override Container<Drawable> Content => content;
-            public AddElementButton()
-            {
-                AddInternal(content);
-                Padding = new MarginPadding(7);
-                RelativeSizeAxes = Axes.Y;
-                Anchor = Anchor.BottomRight;
-                Origin = Anchor.BottomRight;
-                AutoSizeAxes = Axes.X;
-                BackgroundColour = Color4.Black;
-                HoverColour = new Color4(55, 55, 55, 255);
-
-                content.Masking = true;
-                content.CornerRadius = 5;
-                content.BorderThickness = 3f;
-                content.BorderColour = Color4.Black;
-            }
-
-            protected override SpriteText CreateText()
-            {
-                var text = base.CreateText();
-
-                text.Colour = Color4.White;
-                text.Font = new FontUsage(size: 25);
-                text.Margin = new MarginPadding { Horizontal = 6.5f };
-
-                return text;
-            }
         }
 
         private class ObjectManagerScrollContainer : BasicScrollContainer
